Makes ItemAnimators more easily extensible.
Widens the constructor visibility of `GeneralItemAnimator` from package to
protected scope, so that it's possible for library users to extend this class.
Widens the visibility of the inner classes in `RefactoredDefaultItemAnimator`
from private to protected scope, so that it's possible for library users to
extend this class and reuse some of the behaviours while changing others.